pyproject.toml
[tool.poetry]
name = "weathertelegrambot"
version = "0.1.0"
description = ""
authors = ["Bunny Buggs\" <skeree2104@gmail.com\">"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
aiogram = "^2.24"
python-dotenv = "^0.21.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
